requirements.txt
streamlit>=1.28.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0